requirements.txt
langgraph
langchain
langsmith # for LangSmith integration
pydantic
grandalf # for graph visualization in the terminal
docling # for PDF to Markdown conversion
numpy<2
pymupdf # for PDF pre-processing
python-dotenv # for environment variable management
opencv-python # OpenCV for image processing (lapalacian operator)
textstat # for text statistics (readability scores)
langchain_ollama # ollama chatmodel for local llms
langchain_openai # OpenAI chatmodel 
faiss-cpu # vectorstore
langchain-community # for other integrations like FAISS
feedparser # for RSS feed parsing (Arxiv)
tqdm # progressbar 
# Additional:
# python -m spacy download en_core_web_md
# python -m spacy download en_core_web_sm